Login y Register

+++ delete.php
@@ -1,6 +1,11 @@
 <?php 
 require 'database.php';
 
+if (empty($_SESSION["user"])) {
+    header("Location: index.php");
+    return;
+};
+
 // var_dump($_GET["id"]);
 
 $stmt = $conn->prepare("DELETE FROM contacts WHERE id= :idContact");
